Removing ResourceRequestsTrait from ProcessCategoryController

+++ ProcessMaker/Http/Controllers/Api/ProcessCategoryController.php
@@ -10,7 +10,6 @@ use ProcessMaker\Http\Resources\ProcessCategory as Resource;
 
 class ProcessCategoryController extends Controller
 {
-    use ResourceRequestsTrait;
     /**
      * Display a listing of the Process Categories.
      *
@@ -63,7 +62,7 @@ class ProcessCategoryController extends Controller
             $request->input('order_by', 'name'),
             $request->input('order_direction', 'asc')
         );
-        $include = $this->getRequestInclude($request);
+        $include  = $request->input('include') ? explode(',',$request->input('include')) : [];
         $query->with($include);
         $response = $query->paginate($request->input('per_page', 10));
         return new ApiCollection($response);
@@ -126,9 +125,9 @@ class ProcessCategoryController extends Controller
      */
     public function store(Request $request)
     {
+        $request->validate(ProcessCategory::rules());
         $category = new ProcessCategory();
         $category->fill($request->json()->all());
-        $this->validateModel($category, ProcessCategory::rules($category));
         $category->saveOrFail();
         return new Resource($category);
     }
@@ -167,9 +166,8 @@ class ProcessCategoryController extends Controller
      */
     public function update(Request $request, ProcessCategory $processCategory)
     {
+        $request->validate(ProcessCategory::rules($processCategory));
         $processCategory->fill($request->json()->all());
-        //validate model trait
-        $this->validateModel($processCategory, ProcessCategory::rules($processCategory));
         $processCategory->saveOrFail();
         return new Resource($processCategory);
     }
@@ -204,9 +202,12 @@ class ProcessCategoryController extends Controller
      */
     public function destroy(ProcessCategory $processCategory)
     {
-        $this->validateModel($processCategory, [
-            'processes' => 'empty'
-        ]);
+        if ($processCategory->processes->count() !== 0) {
+            return response (
+                ['message'=>'The item should not have associated processes',
+                    'errors'=> ['processes' => $processCategory->processes->count()]],
+                    422);
+        }
 
         $processCategory->delete();
         return response('', 204);
